+++ src/App.jsx
@@ -4,7 +4,7 @@ import postsFromServer from './api/posts.json';
 import commentsFromServer from './api/comments.json';
 import usersFromServer from './api/users.json';
 
-import { PostList } from './components/PostList';
+import { PostList } from './components/PostList/PostList';
 
 const preparedPosts = postsFromServer.map(post => {
   const user = usersFromServer.find(u => u.id === post.userId);
