+++ src/components/CommentList/CommentList.jsx
@@ -1,4 +1,5 @@
 import { CommentInfo } from '../CommentInfo';
+import './CommentList.scss';
 
 export const CommentList = ({ comments }) => (
   <div className="CommentList">
