+++ src/components/CommentList/CommentList.scss
@@ -1,1 +1,7 @@
-// add styles here
+.CommentList {
+  display: flex;
+  flex-direction: column;
+  gap: 0.7em;
+  background-color: #eee;
+  padding: 1em;
+}
