+++ src/components/PostInfo/PostInfo.jsx
@@ -1,5 +1,6 @@
-import { UserInfo } from '../UserInfo';
+import { UserInfo } from '../UserInfo/UserInfo';
 import { CommentList } from '../CommentList/CommentList';
+import './PostInfo.scss';
 
 export const PostInfo = ({ post }) => {
   const { title, body, user, comments } = post;
