+++ src/components/UserInfo/UserInfo.jsx
@@ -1,3 +1,5 @@
+import './UserInfo.scss';
+
 export const UserInfo = ({ user }) => (
   <a className="UserInfo" href={`mailto:${user.email}`}>
     {user.name}
