+++ src/components/UserInfo/UserInfo.scss
@@ -1,1 +1,3 @@
-// add styles here
+.UserInfo {
+  font-weight: bold;
+}
